pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdm-metrics-dashboard"
version = "0.1.0"
description = "SDM metrics dashboard for velocity, quality, and health tracking"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "APScheduler>=3.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
